Revert "파일 복구"
This reverts commit d597397c55e56d4bc2dde8c83349ae40d98fd0b9.

+++ RGM/RGM/Modes/Pluses/Silent/Silent.cs
@@ -14,6 +14,7 @@ using PlayerRoles;
 using Exiled.Events.EventArgs.Player;
 using RGM.API.Features;
 using static RGM.Variables.ServerManagers;
+using VoiceChat.Codec;
 
 namespace RGM.Modes
 {
@@ -21,34 +22,50 @@ namespace RGM.Modes
     public class Silent : Mode
     {
         public override string Name => "사일런트";
-        public override string Description => "쉿!";
+        public override string Description => "쉿! 조용히 이야기하세요!";
         public override string Detail =>
 """
-마이크(Q)를 사용할 수 없습니다.
-채팅(.ㅊ)도 사용할 수 없습니다.
+조용히 이야기해야 합니다. (0.2보다 작게)
+채팅(.ㅊ)은 사용할 수 없습니다.
 """;
         public override string Color => "9E82F5";
 
         public static Silent Instance;
 
+        private OpusDecoder _decoder;
+        private float[] _decodedBuffer;
+
         public override void OnEnabled()
         {
+            _decoder = new OpusDecoder();
+            _decodedBuffer = new float[24000];
+
             Exiled.Events.Handlers.Player.VoiceChatting += OnVoiceChatting;
+        }
+
+        private float GetLoudness(byte[] encodedBuffer, int length)
+        {
+            var len = _decoder.Decode(encodedBuffer, length, _decodedBuffer);
+
+            return len < 0 ? 0f : _decodedBuffer.Max();
         }
 
         public void OnVoiceChatting(VoiceChattingEventArgs ev)
         {
             if (ev.Player.IsAlive)
             {
-                var g = (ExplosiveGrenade)Item.Create(ItemType.GrenadeHE, ev.Player);
-                g.FuseTime = 0f;
-                g.MaxRadius = 0;
-                g.SpawnActive(ev.Player.Position, null);
+                if (GetLoudness(ev.VoiceMessage.Data, ev.VoiceMessage.DataLength) > 0.2)
+                {
+                    var g = (ExplosiveGrenade)Item.Create(ItemType.GrenadeHE, ev.Player);
+                    g.FuseTime = 0f;
+                    g.MaxRadius = 0;
+                    g.SpawnActive(ev.Player.Position, null);
 
-                if (GodModePlayers.Contains(ev.Player))
-                    GodModePlayers.Remove(ev.Player);
+                    if (GodModePlayers.Contains(ev.Player))
+                        GodModePlayers.Remove(ev.Player);
 
-                ev.Player.Kill("입이 근질거리는 것을 참지 못했습니다.");
+                    ev.Player.Kill("입이 근질거리는 것을 참지 못했습니다.");
+                }
             }
         }
     }
